Don't connect to an unknown database cache version.
This commit makes use of the config file added in 75f1270.  If the
database cache has a different version than the one we expect, we cannot
use it anymore.  This commit also adds an easy way to delete the database
cache and config file and suggests that option to users in case of a
version mismatch.