Add an option to compare different data sources.
This new option allows users to specify a country code and look at
similarities and differences between cached databases.

This feature was motivated from comparing MaxMind's non-country code A1 to
assignments in the RIR delegation files.  But the feature is universal and
not at all restricted to this use case.