Add a 9th and 33rd hex character to store new range ends.
Now that we store number ranges as start number to next start number, the
maximum allowed next start numbers are 2^32 + 1 and 2^128 + 1 which need
9 and 33 hex characters.

Example: If we wouldn't add that additional hex character, SQLite would
consider 2^8 + 1, or '100000000' in hex, to be smaller than 2^8, or
'ffffffff' in hex.  But if we write 2^8 as hex '0ffffffff', that problem
is gone.

This change requires moving to new database cache version 0.0.4.