Fix Python 2.6 support
In #69, explicit decoding of the lines was added.  However, this was done using
a keyword argument that isn't supported in Python 2.6.

The good news is that there's no reason to use a keyword argument here - all
versions take the same parameters.  So if we just remove it, all is good.