Update urls; bump version to 4.0.1
The functional URLs in `block_finder/blockfinder.py` now use HTTPS.
MaxMind's GeoIP database URLs are not functional and GeoIP support is
currently broken. A corresponding TODO item was added with a suggestion
for a possible solution. The ISO codes now load a cached version of the
very old file from archive.org over HTTPS. The `-m` option now exits
with an error code of 2.

The codebase has been converted with `2to3-2.7` and a small number of
manual fixes have been applied for Python 3. The GeoIP export features
need to be refactored and are currently non-functional.